Implement solution of PuLP linear programming problem (MPS-13)
- requires moving from python:3.6-alpine3.7 Docker image to python:3.6-slim-stretch, because the default solver used by PuLP on Linux seems not to work well with musl libc — it required glibc. This isnt included in Alpine Linux, which is based on musl

-

+++ app/models.py
@@ -3203,7 +3203,7 @@ class MatchingAttempt(db.Model):
     # MATCHING OUTCOME
 
     # value of objective function, if match was successful
-    score = db.Column(db.Integer())
+    score = db.Column(db.Numeric())
 
 
 # ############################
